Extract 3 largest inner classes from NonCachingTextRenderer (#514)
Extract Glyph, GlyphProducer, and Pipelined_QuadRenderer inner classes
into top-level package-private files:
- TextRendererGlyph.java (226 lines)
- TextRendererGlyphProducer.java (181 lines)
- TextRendererQuadRenderer.java (176 lines)

Each extracted class receives a NonCachingTextRenderer parameter (last
in constructor signature) stored as 'textRenderer' field, replacing
the implicit enclosing-instance reference.

NonCachingTextRenderer reduced from 1842 to 1318 lines (under 1350
target). Widened 14 members from private to package-private to support
cross-class access. Pipelined_QuadRenderer.draw() widened from private
to package-private for flushGlyphPipeline() access.
